Change distinct type to Int64 in columns metadata

+++ noni/Models/ColumnMetadata.cs
@@ -4,7 +4,7 @@ namespace noni.Models {
     // https://stackoverflow.com/questions/58074304/is-polymorphic-deserialization-possible-in-system-text-json
     public class ColumnMetadata
     {
-        public Decimal distinct{get;set;}
+        public Int64 distinct{get;set;}
     }
 
 }
